major revision for win10 variant

+++ raw-install.cmd
@@ -2,7 +2,9 @@
 :: https://github.com/readbeyond/aeneas
 @echo off
 call :check_Permissions
-
+ver | find /i "version 10." > nul
+if %errorlevel%==0 set VERSIONWINDOWS=Windows 10
+if %errorlevel%==0 echo Using Win 10 instal. ExecutionPolicy set to Bypass
 @echo.
 
 echo This installer method uses Chocolatey package manager for Windows
@@ -15,23 +17,24 @@ if "%errorlevel%" == "0" (
   set errorlevel=0
   )
   @echo.
-call where /Q conda.exe
-if "%errorlevel%" == "0" (
-  @echo Miniconda is already installed
-  ) else (
-  @echo   installing Miniconda
-  choco install miniconda
-  set errorlevel=0
-  )
-  @echo.
-choco install ffmpeg  -y
-choco install espeak -s "%cd%" -y
-choco install miniconda -y
-conda install pip -y --no-deps
-choco install vcpluspluspython27 -s "%cd%" -y
-pip install BeautifulSoup
-pip install lxml
-conda install numpy -y --no-deps
+if "%VERSIONWINDOWS%" == "Windows 10" (
+  choco install ffmpeg-w10 -s "%cd%" -y
+  choco install espeak-w10 -s "%cd%" -y
+  choco install miniconda-w10 -s "%cd%" -y
+  choco install vcpluspluspython27-w10 -s "%cd%" -y
+) else (
+  choco install ffmpeg  -y
+  choco install espeak -s "%cd%" -y
+  choco install miniconda -y
+  choco install vcpluspluspython27 -s "%cd%" -y
+)
+rem conda install pip -y --no-deps
+C:\tools\Miniconda\Scripts\pip install BeautifulSoup
+C:\tools\Miniconda\Scripts\pip install lxml
+C:\tools\Miniconda\Scripts\conda install numpy -y --no-deps
+
+echo all done installing prerequesites
+goto :eof
 
 :check_Permissions
     rem from Ben Hooper on http://stackoverflow.com/questions/4051883/batch-script-how-to-check-for-admin-rights
